requirements.txt
openai
python-dotenv
pydantic>=2.7
langchain>=0.2.5
langchain-core>=0.2.8
langgraph>=0.2.0
langsmith>=0.1.90
langchain-openai>=0.1.0
google-search-results
requests
fastapi
uvicorn
supabase